+++ amazon/core/src/main/kotlin/org/http4k/format/autoMapping.kt
@@ -1,6 +1,7 @@
 package org.http4k.format
 
 import dev.forkhandles.values.Value
+import dev.forkhandles.values.ValueFactory
 import org.http4k.connect.amazon.model.ARN
 import org.http4k.connect.amazon.model.AccessKeyId
 import org.http4k.connect.amazon.model.AwsAccount
@@ -13,21 +14,18 @@ import org.http4k.connect.amazon.model.SessionToken
 import org.http4k.connect.amazon.model.Timestamp
 import org.http4k.lens.BiDiMapping
 
-inline fun <reified V : Value<String>, T> AutoMappingConfiguration<T>.text(noinline fn: (String) -> V) =
-    text(BiDiMapping(fn) { it.value })
-
-inline fun <reified V : Value<Long>, T> AutoMappingConfiguration<T>.long(noinline fn: (Long) -> V) =
-    long(BiDiMapping(fn) { it.value })
+inline fun <BUILD, reified VALUE : Value<T>, T : Any> AutoMappingConfiguration<BUILD>.value(fn: ValueFactory<VALUE, T>) =
+    text(fn::parse, fn::print)
 
 fun <T> AutoMappingConfiguration<T>.withAwsCoreMappings() = apply {
-    text(AccessKeyId::of)
-    text(ARN::of)
-    text(AwsService::of)
-    text(AwsAccount::parse)
-    text(Base64Blob::of)
-    text(KMSKeyId::of)
-    text(Region::of)
-    text(SecretAccessKey::of)
-    text(SessionToken::of)
+    value(AccessKeyId)
+    value(ARN)
+    value(AwsService)
+    value(AwsAccount)
+    value(Base64Blob)
+    value(KMSKeyId)
+    value(Region)
+    value(SecretAccessKey)
+    value(SessionToken)
     double(BiDiMapping({ Timestamp.of(it.toLong()) }, { it.value.toDouble() }))
 }