+++ amazon/systemsmanager/client/src/main/kotlin/org/http4k/connect/amazon/systemsmanager/SystemsManagerMoshi.kt
@@ -13,7 +13,7 @@ import org.http4k.format.AwsJsonAdapterFactory
 import org.http4k.format.ConfigurableMoshi
 import org.http4k.format.adapter
 import org.http4k.format.asConfigurable
-import org.http4k.format.text
+import org.http4k.format.value
 import org.http4k.format.withAwsCoreMappings
 import org.http4k.format.withStandardMappings
 
@@ -28,7 +28,7 @@ object SystemsManagerMoshi : ConfigurableMoshi(
 )
 
 fun <T> AutoMappingConfiguration<T>.withSystemsManagerMappings() = apply {
-    text(SSMParameterName::of)
+    value(SSMParameterName)
 }
 
 object SystemsManagerJsonAdapterFactory : AwsJsonAdapterFactory(