+++ CHANGELOG.md
@@ -6,6 +6,7 @@ changes with their rationale when appropriate. Given version `A.B.C.D`, breaking
 ### v2.10.0.0 (uncut)
 - **http4k-connect-amazon-** : Add convenience functions to create clients from the system environment. 
 - **http4k-connect-amazon-** : Removed unused Payload type for various clients.
+- **http4k-connect-*** : Upgrade Values4k
 
 ### v2.9.1.0
 - **http4k-connect-amazon-** : Expose Moshi to client API users for JSON-based systems